check: disable exhaustive checks in test files
It's not quite clear how to make it work. In cases where a package both
defines tests and a sum type, you wind up with multiple distinct
definitions of types that do not satisfy types.Identical, which in turn
results in inexhaustive errors. It's not quite clear how to fix this.