decl: Support //gosumtype:decl for Go 1.19
gofmt in Go 1.19 formats '//go-sumtype:decl' as '// go-sumtype:decl'
because '-' is not allowed in the tool name part of a directive comment,
breaking all sum type declarations after files are formatted.

To keep the tool working with Go 1.19, we have to support an alternative
form of declaration. We choose '//gosumtype:decl' by dropping '-' from
the tool name.